requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# LLM & Embeddings
ollama==0.1.6
chromadb==0.4.18
sentence-transformers==2.2.2

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
httpx==0.25.2
